requirements.txt
requests>=2.28
openai>=1.0
beautifulsoup4>=4.12
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
python-dotenv>=1.0
